fix reiew comments issue
